src/requirements.txt
PyYaml
slackclient
